Update imports and dependencies after consolidation
