+++ task.md
@@ -18,8 +18,8 @@ How we are building it:
 - [x] Remove placeholder transcription job routes until actual implementation needed
 - [x] Consolidate duplicate dev dependencies in pyproject.toml
 - [x] Remove root-level test files (keep only tests/ directory structure)
-- [ ] Simplify TypeScript configuration (single tsconfig.json)
-- [ ] Update imports and dependencies after consolidation
+- [x] Simplify TypeScript configuration (single tsconfig.json)
+- [x] Update imports and dependencies after consolidation
 - [ ] Test that YouTube API functionality still works after simplification
 - [ ] User test: Verify backend starts and can fetch YouTube channel data
